Add tag feature for messages
To tag a message, the message must be of a type SlmMail\Mail\Message\Postmark
instead of Zend\Mail\Message. The new message object has a setter and getter
for the tag. The Postmark transport reads this tag, and doesn't add the tag
when the message is of a Zend\Mail\Message type.