Create explicit exception messages
Name the service inside the exception message. The exception class
is used between different services, so name the service to
understand the service where it came from (otherwise, the stack
trace is the only option).

Also, remove the exception messages where only the response of the
api call is used. Give the message more context to use a fixed part
and a variable response part in the message.