Filter mime parts on type and disposition
Any text/plain and text/html which is marked with a disposition
of attachment should be handled as an attachment.

It is not possible to filter only on disposition, because many
users might set a pdf, png or whatever file to the part and not
marking it as attachment. They expect them to be send as attachment
(see even the example in the manual,
http://zf2.readthedocs.org/en/latest/modules/zend.mail.message.html).

Thus, filter on both type and disposition.